cel definitions moved into a separate package outside of authorizers
